Fixed Spotlight issue, Fixed sound issue on server when vehicles registered on other clients

+++ server/ELS-Server/Class1.cs
@@ -55,7 +55,7 @@ namespace ELS_Server
                 TriggerClientEvent("ELS:VehicleEntered", veh);
             });
             EventHandlers["ELS:FullSync:Unicast"] += new Action(() => { });
-            EventHandlers["ELS:FullSync:Broadcast"] += new Action<System.Dynamic.ExpandoObject, int>((dataDic, playerID) =>
+            EventHandlers["ELS:FullSync:Broadcast"] += new Action<System.Dynamic.ExpandoObject, Int16>((dataDic, playerID) =>
              {
                  var dd = (IDictionary<string, object>)dataDic;
 #if DEBUG
